+++ README.md
@@ -1,0 +1,5 @@
+
+Run:
+```
+go run .
+```
