+++ go.mod
@@ -1,17 +1,17 @@
 // go.mod
 
-module 01-Authorization-RS256
+module log-collection-api
 
 go 1.21
 
 require (
 	github.com/auth0/go-jwt-middleware/v2 v2.2.0
 	github.com/joho/godotenv v1.5.1
+	github.com/rs/cors v1.10.1
+	github.com/unrolled/secure v1.13.0
 )
 
 require (
-	github.com/rs/cors v1.10.1 // indirect
-	github.com/unrolled/secure v1.13.0 // indirect
 	golang.org/x/crypto v0.4.0 // indirect
 	golang.org/x/sync v0.5.0 // indirect
 	gopkg.in/go-jose/go-jose.v2 v2.6.1 // indirect
